+++ src/grandorgue/SettingsOption.h
@@ -22,7 +22,7 @@
 #ifndef SETTINGSOPTION_H
 #define SETTINGSOPTION_H
 
-#include <wx/scrolwin.h>
+#include <wx/panel.h>
 
 enum class GOInitialLoadType;
 template<class T> class GOrgueChoice;
@@ -32,7 +32,7 @@ class wxChoice;
 class wxDirPickerCtrl;
 class wxSpinCtrl;
 
-class SettingsOption : public wxScrolledWindow
+class SettingsOption : public wxPanel
 {
 	enum {
 		ID_WAVE_FORMAT = 200,
@@ -54,8 +54,6 @@ class SettingsOption : public wxScrolledWindow
 		ID_RELEASE_LOAD,
 		ID_CHANNELS,
 		ID_INTERPOLATION,
-		ID_SAMPLE_RATE,
-		ID_SAMPLES_PER_BUFFER,
 		ID_MEMORY_LIMIT,
 		ID_ODF_CHECK,
 		ID_RECORD_DOWNMIX,
@@ -84,8 +82,6 @@ private:
 	wxChoice* m_ReleaseLoad;
 	wxChoice* m_Channels;
 	wxChoice* m_Interpolation;
-	wxChoice* m_SampleRate;
-	wxSpinCtrl* m_SamplesPerBuffer;
 	wxSpinCtrl* m_MemoryLimit;
 	wxChoice* m_Language;
 	unsigned m_OldChannels;